♻ refactor(I18n): The I18n settings currently only set the UICulture
