+++ src/Masa.Blazor/I18n/I18nServiceCollectionExtensions.cs
@@ -10,7 +10,7 @@ public static class I18nServiceCollectionExtensions
 
     internal static IServiceCollection AddI18n(this IServiceCollection services)
     {
-        services.TryAddScoped<I18n>();
+        services.TryAddScoped<Masa.Blazor.I18n>();
 
         return services;
     }
