+++ src/Masa.Blazor/Services/Locale.cs
@@ -2,23 +2,53 @@
 
 public class Locale
 {
-    public string? Current { get; }
+    public CultureInfo UICulture { get; }
 
-    public string? Fallback { get; }
-
-    public string? UICurrent { get; set; }
-
-    public string? UIFallback { get; set; }
-
-    public Locale(string current, string fallback)
+    public Locale(string current, string fallback = "en-US")
     {
-        Current = current;
-        Fallback = fallback;
+        UICulture = GetValidCulture(current, fallback);
     }
 
-    public Locale(string current, string uiCurrent, string fallback, string uiFallback) : this(current, fallback: fallback)
+    public Locale(CultureInfo uiCulture)
     {
-        UICurrent = uiCurrent;
-        UIFallback = uiFallback;
+        UICulture = uiCulture;
+    }
+
+    private static CultureInfo GetValidCulture(string? cultureName, string fallbackCultureName)
+    {
+        CultureInfo? culture = null;
+
+        try
+        {
+            culture = CultureInfo.CreateSpecificCulture(cultureName ?? fallbackCultureName);
+        }
+        catch (Exception)
+        {
+            // ignored
+        }
+
+        if (culture is null && cultureName is not null)
+        {
+            try
+            {
+                culture = CultureInfo.CreateSpecificCulture(fallbackCultureName);
+            }
+            catch (Exception)
+            {
+                // ignored
+            }
+        }
+
+        culture ??= CultureInfo.CurrentCulture;
+
+        // https://github.com/dotnet/runtime/issues/18998#issuecomment-254565364
+        // `CultureInfo.CreateSpecificCulture` has the different behavior in different OS,
+        // so need to standardize the culture.
+        return culture.Name switch
+        {
+            "zh-Hans-CN" => new CultureInfo("zh-CN"),
+            "zh-Hant-CN" => new CultureInfo("zh-TW"),
+            _ => culture
+        };
     }
 }
